requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-chess>=1.999
onnxruntime>=1.17.0
numpy>=1.26.0
